test: TotalQuarterCalculateService 테스트코드
